Closes #9, creators are now pulled in along with authors.
Also fixes an issue where subjects were sometimes counted as contributing people - this is a bt of a conflict with how dublin core deals with contributors - they can either be people or topics for the contribution. Right now, we filter them to be only people (though we should look into if this is the most effective way to do this in the future).